requirements.txt
aiogram==2.25.2
#aiohttp==3.8.6
gspread==5.12.4
google-api-python-client==2.123.0
oauth2client==4.1.3
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
payze==1.0.2
openai==1.30.1
huggingface_hub==0.23.3